Change al_get_opengl_version to return a packed integer.
This is consistent with the other version querying functions, and
eliminates the common mistake of comparing the float result with a
double literal.

cherry-pick https://alleg.svn.sourceforge.net/svnroot/alleg/allegro/branches/5.1@13957

From: Peter Wang <tjaden@example.org>

+++ include/allegro5/allegro_opengl.h
@@ -110,7 +110,7 @@
  */
 
 
-AL_FUNC(float,                 al_get_opengl_version,            (void));
+AL_FUNC(uint32_t,              al_get_opengl_version,            (void));
 AL_FUNC(int,                   al_have_opengl_extension,         (const char *extension));
 AL_FUNC(void*,                 al_get_opengl_proc_address,       (const char *name));
 AL_FUNC(ALLEGRO_OGL_EXT_LIST*, al_get_opengl_extension_list,     (void));
